+++ projects/02_architecture_comparison_lab/viper/Sources/PFVIPERStudyPlannerPresenter.h
@@ -1,0 +1,26 @@
+#import <Foundation/Foundation.h>
+#import "PFVIPERStudyPlannerInteractor.h"
+#import "PFVIPERStudyPlannerRouter.h"
+#import "PFVIPERStudyPlannerView.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface PFVIPERStudyPlannerPresenter : NSObject
+@property (nonatomic, weak) id<PFVIPERStudyPlannerView> view;
+@property (nonatomic, strong) PFVIPERStudyPlannerInteractor *interactor;
+@property (nonatomic, strong) PFVIPERStudyPlannerRouter *router;
+
+- (instancetype)initWithView:(id<PFVIPERStudyPlannerView>)view
+                  interactor:(PFVIPERStudyPlannerInteractor *)interactor
+                      router:(PFVIPERStudyPlannerRouter *)router;
+- (void)bootstrap;
+- (void)addTaskWithTitle:(NSString *)title
+                   notes:(NSString *)notes
+                    tags:(NSArray<NSString *> *)tags
+        estimatedMinutes:(NSInteger)estimatedMinutes;
+- (void)markTaskDoneAtIndex:(NSUInteger)index;
+- (void)showTasksFilteredByTag:(nullable NSString *)tag;
+- (void)saveAndReload;
+@end
+
+NS_ASSUME_NONNULL_END