+++ projects/02_architecture_comparison_lab/viper/Sources/PFVIPERStudyPlannerView.h
@@ -1,0 +1,13 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol PFVIPERStudyPlannerView <NSObject>
+- (void)displayLines:(NSArray<NSString *> *)lines title:(NSString *)title;
+- (void)displayActionMessage:(NSString *)message;
+@end
+
+@interface PFVIPERConsoleView : NSObject <PFVIPERStudyPlannerView>
+@end
+
+NS_ASSUME_NONNULL_END