+++ projects/02_architecture_comparison_lab/viper/Sources/PFVIPERStudyPlannerView.m
@@ -1,0 +1,16 @@
+#import "PFVIPERStudyPlannerView.h"
+
+@implementation PFVIPERConsoleView
+
+- (void)displayLines:(NSArray<NSString *> *)lines title:(NSString *)title {
+    NSLog(@"[VIPER] %@", title);
+    for (NSString *line in lines) {
+        NSLog(@"%@", line);
+    }
+}
+
+- (void)displayActionMessage:(NSString *)message {
+    NSLog(@"%@", message);
+}
+
+@end