+++ projects/02_architecture_comparison_lab/viper/Sources/PFVIPERStudyTaskStore.h
@@ -1,0 +1,12 @@
+#import <Foundation/Foundation.h>
+#import "PFVIPERStudyTask.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface PFVIPERStudyTaskStore : NSObject
+- (NSArray<PFVIPERStudyTask *> *)defaultTasks;
+- (nullable NSString *)saveTasks:(NSArray<PFVIPERStudyTask *> *)tasks error:(NSError **)error;
+- (nullable NSArray<PFVIPERStudyTask *> *)loadTasksFromPath:(NSString *)path error:(NSError **)error;
+@end
+
+NS_ASSUME_NONNULL_END